YUMMY_DIARY/backend/requirements.txt
asgiref==3.2.7
astroid==2.3.3
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.5
django-cors-headers==3.2.1
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
drf-yasg==1.17.1
idna==2.9
inflection==0.4.0
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.2
joblib==0.14.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mysqlclient==1.4.6
numpy==1.17.3
packaging==20.3
pandas==0.25.3
PyJWT==1.7.1
pylint==2.4.4
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.3
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.14.0
sqlparse==0.3.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.9
wrapt==1.11.2
